Added lazy loading routes for projects and editor, updated existing routes
